requirements.txt
hikari==2.0.0.dev115
hikari-lightbulb==2.3.1
python-dotenv==0.21.0
